Made various changes based on feedback, most notably with the clients section
