+++ app/careers/CareerPage.css
@@ -8,55 +8,6 @@
   background-size: cover;
 }
 
-/* || Intern Testimonials */
-.interns-testimonials-heading {
-  font-size: 32px;
-  line-height: 1.5;
-}
-
-.interns-testimonials-sub-heading {
-  font-size: 18px;
-  line-height: 1.7;
-}
-
-.interns-testimonials .careers-page-card {
-  background-color: #000 !important;
-  box-shadow: 0 8px 20px #000000 !important;
-}
-
-.clients-section .careers-page-card {
-  box-shadow: 0 6px 10px #000000 !important;
-}
-
-.careers-page-card {
-  border-radius: 10px;
-  transition: transform 0.3s ease-in-out, box-shadow 0.3s ease-in-out;
-}
-
-.careers-page-card:hover {
-  transform: translateY(-10px);
-  box-shadow: 0 8px 30px rgba(0, 0, 0, 0.2);
-}
-
-.careers-page-carousel .carousel-indicators {
-  margin: 0;
-  bottom: -50px;
-}
-
-.careers-page-carousel .carousel-indicators button {
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background-color: #28303b;
-  opacity: 0.2;
-}
-
-.careers-page-carousel .carousel-indicators button.active {
-  background-color: #000 !important;
-  color: #000 !important;
-  opacity: 1;
-}
-
 /* || Apply */
 .apply {
   background-blend-mode: overlay;
@@ -137,9 +88,18 @@ section.clients-section div.carousel-item {
   max-width: 100%;
 }
 
+.clients-grid .section-card {
+  box-shadow: 0 6px 10px #000000 !important;
+}
+
+.clients-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 20px;
+}
+
 @media (min-width: 992px) {
-  .item-col {
-    flex: 0 0 calc(100% / var(--items-per-row));
-    max-width: calc(100% / var(--items-per-row));
+  .clients-grid {
+    grid-template-columns: 1fr 1fr 1fr;
   }
 }
